// src/types/appwrite.d.ts
export enum ReactionsTargetType {
    NOTE = "note",
    COMMENT = "comment"
}
// ...
export type Reactions = Models.Row & {
    targetType: ReactionsTargetType;
    emoji: string;
    createdAt: string;
    targetId: string;
    userId: string;
}